pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<=3.11.6"
replit = "^3.2.4"
Flask = "^2.1.3"
Flask-SQLAlchemy = "^3.0.3"
Flask-WTF = "^1.1.1"
WTForms = "^3.0.1"
email-validator = "^1.1.3"
openai = "^0.26.4"
bs4 = "^0.0.1"
wikipedia = "^1.4.0"
Flask-Login = "^0.6.2"
geopy = "^2.3.0"
pytz = "^2023.2"
tiktoken = "^0.3.3"

[tool.poetry.dev-dependencies]
debugpy = "^1.6.2"
replit-python-lsp-server = {extras = ["yapf", "rope", "pyflakes"], version = "^1.5.9"}
toml = "^0.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"